Remove explicit install instructions, update refs to cert-manager (#9)
The instructions recommended a very old version of cert-manager which is
no longer supported. It might be better for users if they're linked to
the cert-manager installation instructions on our website, which we try
to keep up to date for each version we release.

I also added a link and updated so that "cert-manager" was always
written in the way we like to style it. See
https://cert-manager.io/docs/faq/style/ for details on that.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>